# backend/internal/http/docs/openapi.yaml
openapi: 3.0.3
info:
  title: HRIS LDP API
  version: 1.0.0
  description: Dokumentasi endpoint utama HRIS LDP.
servers:
  - url: /api
    description: API endpoints
  - url: /
    description: Public/root endpoints
...
paths:
...
  /profile:
    get:
      tags: [Profile]
      summary: Data profile page
      security:
        - cookieAuth: []
      responses:
        '200':
          description: OK
        '401':
          description: Unauthenticated
    patch:
      tags: [Profile]
      summary: Update profil dasar
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Updated
        '422':
          description: Validasi gagal
    delete:
      tags: [Profile]
      summary: Hapus data /profile
      security:
        - cookieAuth: []
      responses:
        '200':
          description: OK
        '401':
          description: Unauthenticated
...
  /pelamar/profil:
    get:
      tags: [Pelamar]
      summary: Ambil data /pelamar/profil
      security:
        - cookieAuth: []
      responses:
        '200':
          description: OK
        '401':
          description: Unauthenticated
    post:
      tags: [Pelamar]
      summary: Proses request /pelamar/profil
      security:
        - cookieAuth: []
      responses:
        '200':
          description: OK
        '401':
          description: Unauthenticated
...
  /profile/password:
    put:
      tags: [Profile]
      summary: Perbarui data /profile/password
      security:
        - cookieAuth: []
      responses:
        '200':
          description: OK
        '401':
          description: Unauthenticated
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: hris_session
...